backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# LangChain
langchain==0.3.14
langchain-openai==0.3.0
langchain-pinecone==0.2.0
langchain-community==0.3.14
langchain-text-splitters==0.3.4

# Pinecone
pinecone-client==5.0.1

# OpenAI
openai==1.58.1

# Document parsing
pypdf==5.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
httpx==0.28.1

# Logging
python-json-logger==3.2.1
